requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
jiter==0.6.1
multidict==6.1.0
openai==1.53.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-telegram-bot==21.6
requests==2.32.3
rfc3986==2.0.0
sniffio==1.3.1
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.15.4
google-genai==1.2.0
google-api-core==2.24.2
SQLAlchemy==2.0.38
asyncpg==0.30.0

# for testing
pytest==8.3.4
pytest-asyncio==0.25.3
aiosqlite==0.21.0
pytest-mock==3.14.0
pytest-cov==6.0.0